update
Add ability to edit project, pledge, user info.

Add permissions

Create 'pledges/' endpoint

+++ crowdfunding/projects/models.py
@@ -7,7 +7,7 @@ from django.db import models
 class Project(models.Model):
     title = models.CharField(max_length=200)
     description = models.TextField()
-    goal = models.IntegerField()
+    goal = models.FloatField()
     image = models.URLField()
     is_open = models.BooleanField()
     date_created = models.DateTimeField(auto_now_add=True)
@@ -19,7 +19,7 @@ class Project(models.Model):
 
 
 class Pledge(models.Model):
-    amount = models.IntegerField()
+    amount = models.FloatField()
     comment = models.CharField(max_length=200)
     anonymous = models.BooleanField()
     project = models.ForeignKey(
